requirement.txt
fastapi
pydantic
uvicorn
sqlalchemy
pytz
pymysql
bcrypt
pyjwt
cryptography
redis
httpx
cryptography